refactor: pagination, sortingbtn 이동
